pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notifer-cli"
version = "1.1.0"
description = "Command-line interface for Notifer notification service"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Notifer Team", email = "hello@notifer.io"},
]
keywords = ["notifications", "cli", "pub-sub", "push-notifications"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Topic :: Communications",
    "Topic :: Utilities",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "sseclient-py>=1.8.0",
]

[project.urls]
Homepage = "https://notifer.io"
Documentation = "https://docs.notifer.io"
Repository = "https://github.com/nexolab-projects/notifer-cli"
Issues = "https://github.com/nexolab-projects/notifer-cli/issues"

[project.scripts]
notifer = "notifer_cli.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["notifer_cli*"]
